+++ retryhttp/wait.py
@@ -1,4 +1,3 @@
-import tenacity
 from typing import Tuple, Type, Union, Sequence
 from retryhttp.helpers import (
     get_default_http_status_exceptions,
@@ -7,9 +6,11 @@ from retryhttp.helpers import (
     is_server_error,
     is_rate_limited,
 )
+from tenacity import RetryCallState, wait_exponential, wait_random_exponential
+from tenacity.wait import wait_base
 
 
-class wait_from_header(tenacity.wait.wait_base):
+class wait_from_header(wait_base):
     """Wait strategy that derives the wait value from an HTTP header.
 
     Args:
@@ -22,12 +23,12 @@ class wait_from_header(tenacity.wait.wait_base):
     def __init__(
         self,
         header: str,
-        fallback: tenacity.wait.wait_base,
+        fallback: wait_base,
     ) -> None:
         self.header = header
         self.fallback = fallback
 
-    def __call__(self, retry_state: tenacity.RetryCallState) -> float:
+    def __call__(self, retry_state: RetryCallState) -> float:
         if retry_state.outcome:
             exc = retry_state.outcome.exception()
             if isinstance(exc, get_default_http_status_exceptions()):
@@ -55,34 +56,34 @@ class wait_rate_limited(wait_from_header):
 
     def __init__(
         self,
-        fallback: tenacity.wait.wait_base = tenacity.wait_exponential(),
+        fallback: wait_base = wait_exponential(),
     ) -> None:
         super().__init__(header="Retry-After", fallback=fallback)
 
 
-class wait_http_errors(tenacity.wait.wait_base):
-    """Context-aware wait strategy based on the type of HTTP error.
+class wait_context_aware(wait_base):
+    """Uses a different wait strategy based on the type of HTTP error.
 
     Args:
         wait_server_errors: Wait strategy to use with server errors.
         wait_network_errors: Wait strategy to use with network errors.
         wait_timeouts: Wait strategy to use with timeouts.
         wait_rate_limited: Wait strategy to use with 429 Too Many Requests.
-        server_error_codes: One or more 5xx HTTP status codes to consider for
+        server_error_codes: One or more 5xx HTTP status codes that will trigger
             `wait_server_errors`. If omitted, defaults to:
 
             - 500
             - 502
             - 503
             - 504
-        network_errors: One or more exceptions to consider for `wait_network_errors`.
+        network_errors: One or more exceptions that will trigger `wait_network_errors`.
             If omitted, defaults to:
 
             - httpx.ConnectError
             - httpx.ReadError
             - httpx.WriteError
             - requests.ConnectionError
-        timeouts: One or more exceptions to consider for `wait_timeouts`. If omitted,
+        timeouts: One or more exceptions that will trigger `wait_timeouts`. If omitted,
             defaults to:
 
             - httpx.ConnectTimeout
@@ -93,10 +94,10 @@ class wait_http_errors(tenacity.wait.wait_base):
 
     def __init__(
         self,
-        wait_server_errors: tenacity.wait.wait_base = tenacity.wait_exponential_jitter(),
-        wait_network_errors: tenacity.wait.wait_base = tenacity.wait_exponential(),
-        wait_timeouts: tenacity.wait.wait_base = tenacity.wait_exponential_jitter(),
-        wait_rate_limited: tenacity.wait.wait_base = wait_rate_limited(),
+        wait_server_errors: wait_base = wait_random_exponential(),
+        wait_network_errors: wait_base = wait_exponential(),
+        wait_timeouts: wait_base = wait_random_exponential(),
+        wait_rate_limited: wait_base = wait_rate_limited(),
         server_error_codes: Union[Sequence[int], int] = (
             500,
             502,
@@ -122,7 +123,7 @@ class wait_http_errors(tenacity.wait.wait_base):
         self.network_errors = network_errors
         self.timeouts = timeouts
 
-    def __call__(self, retry_state: tenacity.RetryCallState) -> float:
+    def __call__(self, retry_state: RetryCallState) -> float:
         if retry_state.outcome:
             exc = retry_state.outcome.exception()
             if is_server_error(exc=exc, status_codes=self.server_error_codes):